Add `.clang-format` file for C++ source files
This patch adds the `.clang-format` file which contains a style to be
applied automatically by the clang-format [1] tool to the project's C++
files.

To apply the style to all the project's C++ files, run:

    $ find -iname '*.cpp' -o -iname '*.hpp' -exec clang-format -i '{}' ';'

As of this patch, the project has no C++ files. This patch is part of an
effort to lay the ground for the introduction of C++ sources in
Babeltrace 2.

I chose to set the `PointerAlignment` style setting to `Left` as
clang-format also uses it for references:

    const char* str

    const std::string& str

[1]: https://clang.llvm.org/docs/ClangFormat.html

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I1def3ec2bf309819878aacef80bd5681d4c44e73
Reviewed-on: https://review.lttng.org/c/babeltrace/+/4423