CONTRIBUTING.adoc: document C++ usage
This patch adds a C++ usage section to `CONTRIBUTING.adoc` as a baseline
of rules and recommended practices to write C++ code for Babeltrace 2.

The reminders at the end of the document are some points I could
initially think of, but we'll add more and probably categorize them as
we write C++ code for this project.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I482fd5d4c3870f90bc8f9b7d7122f19bac0e2757
Reviewed-on: https://review.lttng.org/c/babeltrace/+/4424