Fix: bootstrap with autoconf >= 2.70
We want to enable all warnings and treat them as errors when running
autoreconf but we have to disable 'obsolete' because of
AM_PATH_GLIB_2_0. Turns out that autoconf >= 2.70 will drop all other
warnings categories when 'all' is used. So the only way to enable all
categories but 'obsolete' is to list them all.

Change-Id: I6fae9c86130d79a1bc9d3bb7d0610a24d734532a
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>