src.ctf.fs: add status conversion function
The line changed, in bt_message_iterator_class_initialize_method_status,
converts a ctf_msg_iter_medium_status to a
bt_message_iterator_class_initialize_method_status using a direct cast.
However, there is not a 1:1 mapping between the two enums.  There are
values in the source that are not valid in the destination.

Replace the int cast with a dedicated conversion method.

Change-Id: Iad24a87bb3926ea9191ca00e18e13df933807f65
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/7094
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>