ctf/visitor-generate-ir: rename struct ctf to ctf_visitor_generate_ir
This code uses a different name for an externally visible struct type
(struct ctf_visitor_generate_ir) that is never defined, and an internal
struct type (struct ctx).  This requires some casts from the external
type to the internal type at the entry points.  This seems pointless to
me.  Change the internal type to have the same name as the externally
visible type.  It is still opaque from the point of view of the users
outside this compilation unit.

Change-Id: I1503b50658bf43e5cb4a60dfb811384ac5a24359
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/7100
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>